Extracted "CurrentUserDropdown" component
